правки к ДЗ по каомментариям наставника
