+++ js/constants.js
@@ -1,0 +1,55 @@
+const PHOTOS_COUNT = 25;
+const AVATARS_COUNT = 6;
+const LIKES_COUNT = {
+  MIN: 15,
+  MAX: 200
+};
+
+const NAMES = [
+  'Артём',
+  'Добрыня',
+  'Зигмунд',
+  'Люций',
+  'Азамат',
+  'Панкрат',
+  'Анастасия',
+  'Генриетта',
+  'Корнелия',
+  'Мишель',
+  'Ребекка',
+  'Фатима'
+];
+
+const COMMENTS = [
+  'Всё отлично!',
+  'В целом всё неплохо. Но не всё.',
+  'Когда вы делаете фотографию, хорошо бы убирать палец из кадра. В конце концов это просто непрофессионально.',
+  'Моя бабушка случайно чихнула с фотоаппаратом в руках и у неё получилась фотография лучше.',
+  'Я поскользнулся на банановой кожуре и уронил фотоаппарат на кота и у меня получилась фотография лучше.',
+  'Лица у людей на фотке перекошены, как будто их избивают. Как можно было поймать такой неудачный момент?!'
+];
+
+const DESCRIPTION = [
+  'Крутое',
+  'Класcное',
+  'Ужасное',
+  'Красивое',
+  'Необычное',
+  'Изумительное',
+  'Милое',
+  'Страшное',
+  'Размытое',
+  'Цветное',
+  'Любимое',
+  'Интересное',
+  'Удачное',
+  'Неудачное',
+  'Яркое',
+  'Грустное',
+  'Сказочное',
+  'Мрачное',
+  'Памятное',
+  'Откровенное'
+];
+
+export {PHOTOS_COUNT, AVATARS_COUNT, LIKES_COUNT, NAMES, COMMENTS, DESCRIPTION};
